# control/docs/swagger.json
{
    "components": {
        "schemas": {
...
            "main.apiError": {
                "properties": {
                    "message": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "main.createPATReq": {
                "properties": {
                    "expires_in_days": {
                        "type": "integer"
                    },
                    "label": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "main.githubAppDTO": {
                "properties": {
                    "app_id": {
                        "type": "integer"
                    },
                    "callback_url": {
                        "type": "string"
                    },
                    "key_set": {
                        "type": "boolean"
                    },
                    "name": {
                        "type": "string"
                    },
                    "slug": {
                        "type": "string"
                    },
                    "updated_at": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "main.updateProfileReq": {
                "properties": {
                    "first_name": {
                        "type": "string"
                    },
                    "last_name": {
                        "type": "string"
                    },
                    "public_key": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "BearerAuth": {
                "in": "header",
                "name": "Authorization",
                "type": "apiKey"
            }
        }
    },
    "info": {
        "description": "The self-service API: sign in, manage your profile and your personal access tokens, and create and run VMs.\n\nTwo credentials reach these routes. A browser sends the httpOnly `access_token` cookie. A script sends a personal access token as `Authorization: Bearer dpat_...` — mint one under Settings, or with `POST /me/tokens` from a signed-in session.\n\nA personal access token does everything you can do with your own VMs. It can never reach `/api/v1/admin/*`, even when it belongs to an admin, and it cannot manage tokens: those routes take a password-backed session, so a leaked token cannot mint its successor.\n\nAdmin and client-enrollment routes are deliberately absent here. The first is browser-only by design; the second is a protocol between the control plane and its hosts rather than an API.",
        "title": "dummie control API",
        "version": "1.0"
    },
...
    "paths": {
        "/admin/github-app": {
            "delete": {
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                },
                "summary": "Forget the github app",
                "tags": [
                    "admin"
                ]
            },
            "get": {
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/main.githubAppDTO"
                                }
                            }
                        },
                        "description": "OK"
                    }
                },
                "summary": "The github app backing integrations",
                "tags": [
                    "admin"
                ]
            },
            "put": {
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/main.githubAppDTO"
                                }
                            }
                        },
                        "description": "OK"
                    }
                },
                "summary": "Configure the github app",
                "tags": [
                    "admin"
                ]
            }
        },
        "/integrations": {
            "get": {
                "responses": {},
                "summary": "Integrations you own",
                "tags": [
                    "integrations"
                ]
            },
            "post": {
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {},
                "summary": "Create an integration",
                "tags": [
                    "integrations"
                ]
            }
        },
...
        "/integrations/{id}": {
            "delete": {
                "responses": null,
                "summary": "Delete an integration",
                "tags": [
                    "integrations"
                ]
            },
            "get": {
                "responses": {},
                "summary": "One integration",
                "tags": [
                    "integrations"
                ]
            },
            "patch": {
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {},
                "summary": "Update an integration's flags",
                "tags": [
                    "integrations"
                ]
            }
        },
...
        "/integrations/{id}/repos": {
            "put": {
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": null,
                "summary": "Set which repositories an integration covers",
                "tags": [
                    "integrations"
                ]
            }
        },
...
        "/me": {
            "get": {
                "description": "The same shape the admin section shows for a user. Nothing secret is in it: no password hash, and a public key is public.",
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/main.adminUserDTO"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/main.apiError"
                                }
                            }
                        },
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "Read your account",
                "tags": [
                    "me"
                ]
            },
            "put": {
                "description": "Only the display name and the SSH public key are editable. Username, email, role and every allowance are an admin's to set. An empty public key is accepted and clears it; VMs can still be created without one, they are only reachable through the web console.",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "oneOf": [
                                    {
                                        "type": "object"
                                    },
                                    {
                                        "$ref": "#/components/schemas/main.updateProfileReq",
                                        "summary": "body",
                                        "description": "the whole editable surface"
                                    }
                                ]
                            }
                        }
                    },
                    "description": "the whole editable surface",
                    "required": true
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/main.adminUserDTO"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/main.apiError"
                                }
                            }
                        },
                        "description": "a name over 100 characters, or a key that is not a valid public key"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/main.apiError"
                                }
                            }
                        },
                        "description": "Unauthorized"
                    },
                    "409": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/main.apiError"
                                }
                            }
                        },
                        "description": "that public key is already on another account"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "Update your account",
                "tags": [
                    "me"
                ]
            }
        },
        "/me/tokens": {
            "get": {
                "description": "Only the prefix of each token is returned -- enough to recognise one you still hold, useless to anyone who only has this. Requires a signed-in session: a token cannot enumerate its siblings.",
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/main.tokenList"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/main.apiError"
                                }
                            }
                        },
                        "description": "Unauthorized"
                    },
                    "403": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/main.apiError"
                                }
                            }
                        },
                        "description": "authenticated with a personal access token"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "List your personal access tokens",
                "tags": [
                    "tokens"
                ]
            },
            "post": {
                "description": "The response is the only place the raw token appears; it is stored hashed and cannot be shown again. Omit expires_in_days for a token that never expires. Requires a signed-in session, so a leaked token cannot mint its successor and outlive the revocation meant to end it.",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "oneOf": [
                                    {
                                        "type": "object"
                                    },
                                    {
                                        "$ref": "#/components/schemas/main.createPATReq",
                                        "summary": "body",
                                        "description": "a label is required"
                                    }
                                ]
                            }
                        }
                    },
                    "description": "a label is required",
                    "required": true
                },
                "responses": {
                    "201": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/main.createdTokenResp"
                                }
                            }
                        },
                        "description": "the raw token, shown once"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/main.apiError"
                                }
                            }
                        },
                        "description": "Bad Request"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/main.apiError"
                                }
                            }
                        },
                        "description": "Unauthorized"
                    },
                    "403": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/main.apiError"
                                }
                            }
                        },
                        "description": "authenticated with a personal access token"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "Create a personal access token",
                "tags": [
                    "tokens"
                ]
            }
        },
        "/me/tokens/{id}": {
            "delete": {
                "description": "Anything using it stops working and no record of it is kept. Revoke instead when you want the row to stay.",
                "parameters": [
                    {
                        "description": "token id",
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "format": "uuid",
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "deleted"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/main.apiError"
                                }
                            }
                        },
                        "description": "Bad Request"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/main.apiError"
                                }
                            }
                        },
                        "description": "Unauthorized"
                    },
                    "403": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/main.apiError"
                                }
                            }
                        },
                        "description": "authenticated with a personal access token"
                    },
                    "404": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/main.apiError"
                                }
                            }
                        },
                        "description": "no such token, or it belongs to somebody else"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "Delete a personal access token",
                "tags": [
                    "tokens"
                ]
            }
        },
...
        "/oidc/{slug}/callback": {
            "get": {
                "parameters": [
                    {
                        "description": "provider slug",
                        "in": "path",
                        "name": "slug",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "302": {
                        "description": "Found"
                    }
                },
                "summary": "Complete a federated sign-in",
                "tags": [
                    "auth"
                ]
            }
        },
        "/oidc/{slug}/start": {
            "get": {
                "parameters": [
                    {
                        "description": "provider slug",
                        "in": "path",
                        "name": "slug",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "same-origin path to land on afterwards",
                        "in": "query",
                        "name": "redirect",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "302": {
                        "description": "Found"
                    }
                },
                "summary": "Begin a federated sign-in",
                "tags": [
                    "auth"
                ]
            }
        },
        "/openapi.json": {
            "get": {
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        },
                        "description": "the OpenAPI document this page is rendered from"
                    },
                    "404": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/main.apiError"
                                }
                            }
                        },
                        "description": "no spec was built into this binary"
                    }
                },
                "tags": [
                    "docs"
                ]
            }
        },
...
    },
    "openapi": "3.1.0",
    "servers": [
        {
            "url": "/api/v1"
        }
    ]
}